+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-21T14:43:58.401Z",
+    "2026-02-21T15:10:16.389Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {