+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -201,7 +201,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kj0ah086ypt6ad2c08yebhvk"),
+  _id: z.string().default("conv_01kj0c151096jc59a83g8n8bqq"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),