+++ packages/orq-rc/RELEASES.md
@@ -7079,3 +7079,13 @@ Based on:
 - [typescript v4.4.0-rc.25] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.25 - packages/orq-rc
+
+## 2026-02-21 15:13:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.723.0 (2.835.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.26] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.26 - packages/orq-rc